@@ -11,17 +11,23 @@
             "url": "https://github.com/Lan2Play/dedicated-server-api",
             "no-api": true,
             "only": ["maniaplanet/dedicated-server-api"]
+        },
+        {
+            "type": "github",
+            "url": "https://github.com/laravel-shift/laravel-steam-auth",
+            "no-api": true,
+            "only": ["invisnik/laravel-steam-auth"]
         }
     ],
     "license": "MIT",
     "type": "project",
     "require": {
         "php": "8.2.*",
-        "laravel/framework": "9.52.16",
+        "laravel/framework": "^10.0",
         "laravelcollective/html": "6.4.1",
         "laravel/legacy-factories": "1.3.2",
-        "webonyx/graphql-php": "dev-master",
-        "invisnik/laravel-steam-auth": "4.4.0",
+        "webonyx/graphql-php": "15.7.0",
+        "invisnik/laravel-steam-auth": "dev-l10-compatibility",
         "laravel/socialite": "5.9.1",
         "ignited/laravel-omnipay": "3.5.0",
         "doctrine/dbal": "3.7.0",
@@ -33,7 +39,7 @@
         "laravel/tinker": "2.8.2",
         "scssphp/scssphp": "1.11.1",
         "omnipay/paypal": "3.0.2",
-        "cviebrock/eloquent-sluggable": "9.0.0",
+        "cviebrock/eloquent-sluggable": "10.0.0",
         "barryvdh/laravel-debugbar": "3.9.2",
         "omnipay/stripe": "3.2.0",
         "stripe/stripe-php": "8.12.0",
